third_party/WebKit/LayoutTests/editing/deleting/delete-hr.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable><div>Before</div><hr>a|fter</div',
        selection => {
            selection.document.execCommand('delete');
            selection.document.execCommand('delete');
        },
        '<div contenteditable><div>Before</div>|fter</div>'),
        'Delete HR');
</script>
